+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 6.6.0
+
+- Removed plugin `AfterTransitionIndex`.
+- Added plugin `ApplyConditions`.
+
 # 6.5.3
 
 - Docker support added.
@@ -13,7 +18,7 @@
 # 6.5.0
 
 - After schema deleting all schema data deleting is added.
-- After transition deleteing dispatchers deleting is added.
+- After transition deleting dispatchers deleting is added.
 
 # 6.4.1